Fix overlays
