+++ src/config/theme/basePalette.ts
@@ -1,5 +1,4 @@
 import { Theme } from "@peersyst/react-native-components";
-import { ThemeOverlay, ThemeOverlays } from "config/theme/theme.declarations";
 
 //Base colors
 export const white = "#FFFFFF";
@@ -24,76 +23,6 @@ export const gradient: Theme["palette"]["gradient"] = {
     greenViolet: [green["200"], violet],
 };
 
-//Overlays
-//Overlay900 - #000000
-export const overlay900: ThemeOverlay = {
-    "8%": "#00000014",
-    "12%": "#0000001F",
-    "16%": "#00000029",
-    "24%": "#0000003D",
-    "32%": "#00000052",
-    "48%": "#0000007A",
-    "60%": "#00000099",
-    "72%": "#000000B7",
-    "80%": "#000000CC",
-};
-//Overlay700 - #262626
-export const overlay700: ThemeOverlay = {
-    "8%": "#26262614",
-    "12%": "#2626261F",
-    "16%": "#26262629",
-    "24%": "#2626263D",
-    "32%": "#26262652",
-    "48%": "#2626267A",
-    "60%": "#26262699",
-    "72%": "#262626B7",
-    "80%": "#262626CC",
-};
-//Overlay500 - #292929
-export const overlay500: ThemeOverlay = {
-    "8%": "#29292914",
-    "12%": "#2929291F",
-    "16%": "#29292929",
-    "24%": "#2929293D",
-    "32%": "#29292952",
-    "48%": "#2929297A",
-    "60%": "#29292999",
-    "72%": "#292929B7",
-    "80%": "#292929CC",
-};
-//Overlay300 - #a7a7a7
-export const overlay300: ThemeOverlay = {
-    "8%": "#A7A7A714",
-    "12%": "#A7A7A71F",
-    "16%": "#A7A7A729",
-    "24%": "#A7A7A73D",
-    "32%": "#A7A7A752",
-    "48%": "#A7A7A77A",
-    "60%": "#A7A7A799",
-    "72%": "#A7A7A7B7",
-    "80%": "#A7A7A7CC",
-};
-//Overlay100 - #FFFFFF
-export const overlay100: ThemeOverlay = {
-    "8%": "#FFFFFF14",
-    "12%": "#FFFFFF1F",
-    "16%": "#FFFFFF29",
-    "24%": "#FFFFFF3D",
-    "32%": "#FFFFFF52",
-    "48%": "#FFFFFF7A",
-    "60%": "#FFFFFF99",
-    "72%": "#FFFFFFB8",
-    "80%": "#FFFFFFCC",
-};
-
-export const overlay: ThemeOverlays = {
-    100: overlay100,
-    300: overlay300,
-    500: overlay500,
-    700: overlay700,
-    900: overlay900,
-};
-
 //Base theme
 export const basePalette: Partial<Theme["palette"]> = {
     primary: green["200"],
@@ -109,7 +38,6 @@ export const basePalette: Partial<Theme["palette"]> = {
     green,
     gradient,
     wallet: [violet, green["200"], red, gold, purple, blue, orange],
-    overlay,
 };
 
 export const WALLET_GRADIENT_LENGTH = basePalette?.wallet?.length || 0;
