+++ src/config/theme/darkTheme.ts
@@ -1,6 +1,7 @@
 import { createTheme, Theme } from "@peersyst/react-native-components";
 import { theme } from "./theme";
-import { basePalette, black, blue, green, orange, overlay } from "./basePalette";
+import { basePalette, black, blue, green, orange } from "./basePalette";
+import { ThemeOverlay, ThemeOverlays } from "config/theme/theme.declarations";
 
 //Custom light theme colors
 export const red = "#C21212";
@@ -28,6 +29,76 @@ const gray: Theme["palette"]["gray"] = {
     0: "#1A1A1A",
 };
 
+//Overlays
+//Overlay100 - #000000
+export const overlay100: ThemeOverlay = {
+    "8%": "#00000014",
+    "12%": "#0000001F",
+    "16%": "#00000029",
+    "24%": "#0000003D",
+    "32%": "#00000052",
+    "48%": "#0000007A",
+    "60%": "#00000099",
+    "72%": "#000000B7",
+    "80%": "#000000CC",
+};
+//Overlay300 - #262626
+export const overlay300: ThemeOverlay = {
+    "8%": "#26262614",
+    "12%": "#2626261F",
+    "16%": "#26262629",
+    "24%": "#2626263D",
+    "32%": "#26262652",
+    "48%": "#2626267A",
+    "60%": "#26262699",
+    "72%": "#262626B7",
+    "80%": "#262626CC",
+};
+//Overlay500 - #292929
+export const overlay500: ThemeOverlay = {
+    "8%": "#29292914",
+    "12%": "#2929291F",
+    "16%": "#29292929",
+    "24%": "#2929293D",
+    "32%": "#29292952",
+    "48%": "#2929297A",
+    "60%": "#29292999",
+    "72%": "#292929B7",
+    "80%": "#292929CC",
+};
+//Overlay700 - #a7a7a7
+export const overlay700: ThemeOverlay = {
+    "8%": "#A7A7A714",
+    "12%": "#A7A7A71F",
+    "16%": "#A7A7A729",
+    "24%": "#A7A7A73D",
+    "32%": "#A7A7A752",
+    "48%": "#A7A7A77A",
+    "60%": "#A7A7A799",
+    "72%": "#A7A7A7B7",
+    "80%": "#A7A7A7CC",
+};
+//Overlay900 - #FFFFFF
+export const overlay900: ThemeOverlay = {
+    "8%": "#FFFFFF14",
+    "12%": "#FFFFFF1F",
+    "16%": "#FFFFFF29",
+    "24%": "#FFFFFF3D",
+    "32%": "#FFFFFF52",
+    "48%": "#FFFFFF7A",
+    "60%": "#FFFFFF99",
+    "72%": "#FFFFFFB8",
+    "80%": "#FFFFFFCC",
+};
+
+export const overlay: ThemeOverlays = {
+    100: overlay100,
+    300: overlay300,
+    500: overlay500,
+    700: overlay700,
+    900: overlay900,
+};
+
 //Dark theme
 const darkTheme = createTheme({
     ...theme,
@@ -39,6 +110,7 @@ const darkTheme = createTheme({
         background: black,
         text: gray[900],
         gray,
+        overlay,
         backdrop: overlay["500"]["48%"],
         glass: overlay[100]["32%"],
         component: {
